Changed topdown to iterative, and changed folder structure
Changed the topdown approach to use iteration/looping for the calculation instead of recursion (elegance versus stack overflow for large datasets issue). Added more general structure to the execuction format with ability to choose the algorithm, and the interpolation method. Finally, neatened up the file structure by moving the wrapper, fit, and segment files to the trendhelpers folder